refactor: register Perplexity Search tool with all 8 NAT frameworks
Per maintainer @bbednarski9's request on PR #1903, extend the Perplexity
internet search tool from LangChain-only to all 8 frameworks supported
by LLMFrameworkEnum: LANGCHAIN, LLAMA_INDEX, CREWAI, SEMANTIC_KERNEL,
AGNO, ADK, STRANDS, AUTOGEN.

The tool implementation is HTTP-only (raw httpx calls to
api.perplexity.ai/search) with no framework-specific imports, so it can
live in a single framework-agnostic location and be wrapped for each
target framework via NAT's existing tool_wrapper.py mechanism in each
plugin package.

Changes:
- Move tool from packages/nvidia_nat_langchain/.../tools/ to
  packages/nvidia_nat_core/.../tool/ (framework-agnostic location)
- Expand framework_wrappers list to include all 8 LLMFrameworkEnum values
- Move tests to nvidia_nat_core's test suite
- Update docs to reflect multi-framework support
