+++ packages/nvidia_nat_langchain/src/nat/plugins/langchain/tools/register.py
@@ -20,6 +20,5 @@
 
 from . import code_generation_tool
 from . import exa_internet_search
-from . import perplexity_internet_search
 from . import tavily_internet_search
 from . import wikipedia_search
